requirements.txt
fastapi==0.135.2
uvicorn==0.23.2
redis==7.4.0
pydantic==2.12.5
google-generativeai==0.8.6
faiss-cpu==1.13.2
numpy==2.4.4
